+++ etl/steps/data/garden/democracy/2024-03-07/vdem/vdem_aggregate.py
@@ -708,7 +708,7 @@ def make_table_with_dummies(tb: Table) -> Table:
     indicator_names = [indicator["name_new"] for indicator in indicators]
 
     ## Get dummy indicator table
-    tb_ = cast(Table, pd.get_dummies(tb_, dummy_na=True, columns=indicator_names))
+    tb_ = cast(Table, pd.get_dummies(tb_, dummy_na=True, columns=indicator_names, dtype=int))
 
     ## Add missing metadata to dummy indicators
     dummy_cols = []
